Rewriting the middleware system, step 1
