+++ vk/mid/errors.go
@@ -1,0 +1,1 @@
+package mid
